+++ docs/release-notes.md
@@ -26,6 +26,7 @@ and
 <a href="/docs/snapshot/javadoc/index.html">javadoc</a>
 </p>
 
+- Added `Arbitrary.forEachValue(Consumer action)`
 - Fixed [Kotlin compatibility issue](https://github.com/jlink/jqwik/issues/52)
 - Fixed [decimal generation bug](https://github.com/jlink/jqwik/issues/50)
 - Update to JUnit Platform 5.4.1